fix: 修复浮点数无法绘制成图表的问题

+++ back-end/src/main/java/pers/nek0peko/datas/dto/data/chart/option/BarOptionDTO.java
@@ -10,7 +10,7 @@ import java.util.List;
  * BarOptionDTO
  *
  * @author nek0peko
- * @date 2023/04/20
+ * @date 2023/05/07
  */
 @ApiModel
 @Builder
@@ -54,7 +54,7 @@ public class BarOptionDTO extends ChartOptionDTO {
 
         private String name;
 
-        private List<Integer> data;
+        private List<Float> data;
 
     }
