+++ back-end/src/main/java/pers/nek0peko/datas/dto/data/chart/option/LineOptionDTO.java
@@ -54,7 +54,7 @@ public class LineOptionDTO extends ChartOptionDTO {
 
         private String name;
 
-        private List<Integer> data;
+        private List<Float> data;
 
     }
